+++ xkcd/xkcd.psm1
@@ -1,6 +1,7 @@
 $Public = @( Get-ChildItem -Path "$PSScriptRoot/Public/*.ps1" -Recurse )
+$Private = @( Get-ChildItem -Path "$PSScriptRoot/Private/*.ps1" -Recurse -ErrorAction SilentlyContinue )
 
-@($Public) | ForEach-Object {
+@($Public + $Private) | ForEach-Object {
     try {
         . $_.FullName
     }